+++ docs/changelog.rst
@@ -43,6 +43,7 @@ Changes in 0.8.X
 - Undefined data should not override instance methods (#49)
 - Added Django Group and Permission (#142)
 - Added Doc class and pk to Validation messages (#69)
+- Fixed Documents deleted via a queryset don't call any signals (#105)
 
 Changes in 0.7.9
 ================